fix(numbers): two bugs that were live on main, and a way to test a purchase
Both were found merging feat/virtual-numbers back in, and both were on main.

## The funding reconciler had not settled a top-up since it was written

    PREPARE p(float8) AS SELECT now() - make_interval(hours => $1);
    ERROR: function make_interval(hours => double precision) does not exist

make_interval's signature is `hours integer … secs double precision`. Only
secs is a float. Both sweep queries bound ABANDON_AFTER_HOURS as f64, so
Postgres could not resolve the function and the statement failed at prepare
time — every 30 seconds, for the life of the process.

What that costs: the sweep is what chases a Paystack top-up whose payer
never came back to the browser. It never ran. Those intents were never
verified and never abandoned, so the money is taken and the balance never
moves until somebody notices by hand. Nothing failed loudly; a warn line
every 30s is all it produced. Bound as i32 now, which is what the function
takes. No test covered the sweep, which is why it stayed hidden.

## A country could advertise one place and sell numbers from another

number_countries.code is unique and 5SIM files French Guiana under `fr` —
France's ISO code, not its own `gf`. Of 153 country keys on the live guest
list, 152 codes are distinct; that is the only collision.

Reusing the row already holding the code stopped the sweep aborting on
number_countries_code_key, but the reused row still took the incoming name
and dial code while keeping the other key's provider_country. Driven through
the real pair, the row came out as:

    code=FR name="French Guiana" dial=+594 provider_country=france

number_routes resolves a country by code and buys with the provider_country
on that row, so it offers French Guiana at +594 and hands back French
numbers at +33. Which name landed depended on HashMap iteration order, and
both keys resolving to one country_id meant their stock and cost overwrote
each other on every sweep.

Now the code belongs to one supplier key and the other is skipped — no id
back, so no prices are written against a row it does not own. Lowest key
wins rather than first writer: the sync had already run before any rule
existed and frenchguiana held FR, so France was being skipped permanently,
with no amount of re-running able to move it. The winner takes the row over
in place, keeping the id every existing order points at. The sweep iterates
sorted so the loser never prices a row the winner then takes over.

upsert_country_reuses_iso_when_provider_keys_differ asserted the id and the
provider_country but never the name or dial code, which is how the row got
to disagree with itself while the suite stayed green. Replaced by two tests
that pin the whole row, from both starting states.

## devtools credit

Puts naira in an account nobody paid for, so a purchase can be walked end
to end on staging. Refuses to run unless APP_ENV is staging or development,
and posts as Adjustment rather than NgnDeposit — booking it as a deposit
would put it in front of the funding reconciler as a card charge that never
happened. Every credit carries {"source": "manual-test-credit"}.

+++ backend-rs/crates/api/src/funding_reconciler.rs
@@ -30,7 +30,12 @@ const SETTLE_AFTER_SECONDS: i64 = 20;
 /// After this, a pending intent is abandoned rather than verified forever.
 /// Paystack's checkout link is long dead by then; what remains is a row that
 /// would otherwise be re-verified every 30 seconds for the life of the process.
-const ABANDON_AFTER_HOURS: i64 = 24;
+///
+/// `i32` because it is bound to `make_interval(hours => ...)`, whose `hours`
+/// argument is `integer`. Only its `secs` argument is `double precision`, so
+/// binding hours as a float resolves to no function at all and fails the whole
+/// sweep at runtime rather than at compile time.
+const ABANDON_AFTER_HOURS: i32 = 24;
 
 /// How many to settle per pass. A backlog drains over several sweeps rather than
 /// opening hundreds of provider calls at once.
@@ -73,7 +78,7 @@ async fn sweep(state: &AppState) -> anyhow::Result<()> {
           LIMIT $3",
     )
     .bind(SETTLE_AFTER_SECONDS as f64)
-    .bind(ABANDON_AFTER_HOURS as f64)
+    .bind(ABANDON_AFTER_HOURS)
     .bind(BATCH)
     .fetch_all(&state.db)
     .await?;
@@ -101,7 +106,7 @@ async fn sweep(state: &AppState) -> anyhow::Result<()> {
             AND provider = 'paystack'
             AND created_at < now() - make_interval(hours => $1)",
     )
-    .bind(ABANDON_AFTER_HOURS as f64)
+    .bind(ABANDON_AFTER_HOURS)
     .execute(&state.db)
     .await?
     .rows_affected();
